Cythonised the slowest parts AltSim (should probably just start calling it Sim) in a separate version called AltSim_C. Need to compile the cython by calling make in the csim directory. Resulting simulation is many times faster.

+++ DensityField.py
@@ -4,7 +4,7 @@ import numpy as np
 class DensityField:
     def __init__(self, shape):        
         self._shape = shape
-        self._d = np.zeros((3, *shape), dtype=np.float32)
+        self._d = np.zeros((3, *shape))
         
         self._colours = np.array([np.array([1.0, 0.8, 0.2]), 
                                   np.array([0.5, 1.0, 0.5]),
